r/aws_kms_key: Tidy up eventual consistency checks.
